Extract profile edition to separate function
